+++ api/app/tests/Fixture/DeputySet.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Tests\OPG\Digideps\Backend\Fixture;
 
+use OPG\Digideps\Backend\Domain\Deputy\DeputyType;
+
 final readonly class DeputySet
 {
     /**
@@ -14,5 +16,20 @@ final readonly class DeputySet
     public function __construct(DeputyDescriptor ...$descriptors)
     {
         $this->descriptors = $descriptors;
+    }
+
+    public static function oneLay(string $deputyReference = 'lay1'): DeputySet
+    {
+        return new DeputySet(new DeputyDescriptor($deputyReference, DeputyType::LAY));
+    }
+
+    public static function oneNamedPro(string $deputyReference = 'pro1'): DeputySet
+    {
+        return new DeputySet(new DeputyDescriptor($deputyReference, DeputyType::PRO));
+    }
+
+    public static function oneNamedPa(string $deputyReference = 'pa1'): DeputySet
+    {
+        return new DeputySet(new DeputyDescriptor($deputyReference, DeputyType::PA));
     }
 }
