+++ api/app/tests/Integration/v2/Controller/CourtOrderControllerTest.php
@@ -4,87 +4,37 @@ declare(strict_types=1);
 
 namespace Tests\OPG\Digideps\Backend\Integration\v2\Controller;
 
-use OPG\Digideps\Backend\Domain\CourtOrder\CourtOrderKind;
-use OPG\Digideps\Backend\Domain\CourtOrder\CourtOrderType;
+use OPG\Digideps\Backend\Domain\CourtOrder\CourtOrderReportType;
 use OPG\Digideps\Backend\Domain\Deputy\DeputyType;
-use OPG\Digideps\Backend\Entity\CourtOrder;
-use OPG\Digideps\Backend\Entity\Deputy;
 use OPG\Digideps\Backend\Entity\PreRegistration;
 use OPG\Digideps\Backend\Entity\User;
-use OPG\Digideps\Backend\TestHelpers\ReportTestHelper;
-use Tests\OPG\Digideps\Backend\Behat\v2\Helpers\FixtureHelper;
+use Tests\OPG\Digideps\Backend\Fixture\CourtOrderDescriptor;
+use Tests\OPG\Digideps\Backend\Fixture\DeputySet;
+use Tests\OPG\Digideps\Backend\Fixture\Scenario;
+use Tests\OPG\Digideps\Backend\Fixture\UserType;
 use Tests\OPG\Digideps\Backend\Integration\Controller\AbstractTestController;
-use Tests\OPG\Digideps\Backend\Integration\Controller\JsonHttpTestClient;
 
 class CourtOrderControllerTest extends AbstractTestController
 {
-    private static JsonHttpTestClient $client;
-    private static FixtureHelper $fixtureHelper;
-    private static ReportTestHelper $reportTestHelper;
-
     public static function setUpBeforeClass(): void
     {
         parent::setUpBeforeClass();
-
-        $container = static::getContainer();
-
-        self::$fixtureHelper = $container->get(FixtureHelper::class);
-        self::$reportTestHelper = ReportTestHelper::create();
-        self::$client = new JsonHttpTestClient(self::$frameworkBundleClient, self::$jwtService);
-    }
-
-    private function createDeputyForUser(User $user): Deputy
-    {
-        $deputy = new Deputy();
-        $deputy->setDeputyType(DeputyType::LAY);
-        $deputy->setEmail1($user->getEmail());
-        $deputy->setDeputyUid('748723' . rand(1, 99999));
-        $deputy->setFirstname('name' . time());
-        $deputy->setLastname('surname' . time());
-        $deputy->setOrganisation(null);
-        $deputy->setUser($user);
-
-        self::$fixtures->persist($deputy);
-        self::$fixtures->flush();
-
-        return $deputy;
-    }
-
-    /**
-     * @return array{User, CourtOrder} [$user for $deputy], $courtOrder associated with $deputy]
-     */
-    private function addUserAndCourtOrderAndDeputy($emailAddress): array
-    {
-        // add a court order, and make the user a deputy on it
-        $courtOrder = self::$fixtures->createCourtOrder(substr('' . hexdec(uniqid()), -8), CourtOrderType::PFA, CourtOrderKind::Single, 'ACTIVE');
-
-        $user = self::$fixtures->createUser($emailAddress, User::ROLE_LAY_DEPUTY);
-        self::$fixtureHelper->setPassword($user);
-
-        // associate deputy with court order
-        $deputy = $this->createDeputyForUser($user);
-        $deputy->associateWithCourtOrder($courtOrder);
-
-        self::$fixtures->persist($courtOrder, $user, $deputy)->flush();
-
-        return [$user, $courtOrder];
     }
 
     public function testGetByUidActionNoAuthFail()
     {
-        self::$client->assertEndpointNeedsAuth('GET', '/v2/courtorder/71101111');
+        self::assertEndpointNeedsAuth('GET', '/v2/courtorder/71101111');
     }
 
     public function testGetByUidActionCourtOrderNotFoundFail(): void
     {
-        $user = self::$fixtures->createUser('fail-not-found-court-order-test@opg.gov.uk', User::ROLE_LAY_DEPUTY);
-        self::$fixtureHelper->setPassword($user);
+        $user = self::$fixtureService->instantiateOnlyUser(UserType::Deputy, DeputyType::LAY);
 
         // log in and fetch court order which doesn't exist
-        $token = self::$client->login('fail-not-found-court-order-test@opg.gov.uk', 'DigidepsPass1234', self::$deputySecret);
+        $token = self::loginAsDeputy($user->getEmail());
 
         // make the API call
-        self::$client->assertJsonRequest(
+        self::assertJsonRequest(
             'GET',
             '/v2/courtorder/9292777777',
             ['AuthToken' => $token, 'mustFail' => true, 'assertResponseCode' => 404]
@@ -94,19 +44,16 @@ class CourtOrderControllerTest extends AbstractTestController
     public function testGetByUidActionUserIsNotADeputyFail(): void
     {
         // add a court order
-        $courtOrder = self::$fixtures->createCourtOrder('92954529292', CourtOrderType::HW, CourtOrderKind::Single, 'ACTIVE');
-        self::$fixtures->persist($courtOrder);
-        self::$fixtures->flush();
+        ['orders' => [['hw' => ['order' => $courtOrder]]]] = self::$fixtureService->instantiateScenario(Scenario::newSimpleLayScenario(reportType: CourtOrderReportType::OPG104));
 
         // log in, and fetch court order which exists, but user has no deputy record
-        $user = self::$fixtures->createUser('fail-user-not-deputy-court-order-test@opg.gov.uk', User::ROLE_LAY_DEPUTY);
-        self::$fixtureHelper->setPassword($user);
+        $user = self::$fixtureService->instantiateOnlyUser(UserType::Deputy, DeputyType::LAY);
 
-        $token = self::$client->login('fail-user-not-deputy-court-order-test@opg.gov.uk', 'DigidepsPass1234', self::$deputySecret);
+        $token = self::loginAsDeputy($user->getEmail());
 
-        self::$client->assertJsonRequest(
+        self::assertJsonRequest(
             'GET',
-            '/v2/courtorder/92954529292',
+            "/v2/courtorder/{$courtOrder->getId()}",
             ['AuthToken' => $token, 'mustFail' => true, 'assertResponseCode' => 404]
         );
     }
@@ -114,52 +61,32 @@ class CourtOrderControllerTest extends AbstractTestController
     public function testGetByUidActionUserIsNotADeputyOnCourtOrderFail(): void
     {
         // add a court order
-        $courtOrder = self::$fixtures->createCourtOrder('9292929292', CourtOrderType::HW, CourtOrderKind::Single, 'ACTIVE');
-        self::$fixtures->persist($courtOrder);
-        self::$fixtures->flush();
+        ['orders' => [['hw' => ['order' => $courtOrder]]]] = self::$fixtureService->instantiateScenario(Scenario::newSimpleLayScenario(reportType: CourtOrderReportType::OPG104));
 
         // create a deputy for the user, so they have a valid deputy record, but don't associate with court order
-        $user = self::$fixtures->createUser('fail-not-deputy-on-court-order-test@opg.gov.uk', User::ROLE_LAY_DEPUTY);
-        self::$fixtureHelper->setPassword($user);
-        $this->createDeputyForUser($user);
+        ['persons' => ['users' => ['lay1' => $user]]] = self::$fixtureService->instantiateScenario(Scenario::newSimpleLayScenario());
 
         // log in, and fetch court order which exists but for which the logged-in user is not a deputy
-        $token = self::$client->login('fail-not-deputy-on-court-order-test@opg.gov.uk', 'DigidepsPass1234', self::$deputySecret);
+        $token = self::loginAsDeputy($user->getEmail());
 
-        self::$client->assertJsonRequest(
+        self::assertJsonRequest(
             'GET',
-            '/v2/courtorder/9292929292',
+            "/v2/courtorder/{$courtOrder->getId()}",
             ['AuthToken' => $token, 'mustFail' => true, 'assertResponseCode' => 404]
         );
     }
 
     public function testGetByUidActionSuccess(): void
     {
-        [$user, $courtOrder] = $this->addUserAndCourtOrderAndDeputy('successful-court-order-test@opg.gov.uk');
-
-        // client
-        $client = self::$fixtures->createClient($user);
-        self::$fixtures->persist($client)->flush();
-
-        // add an unsubmitted (current) report to the court order
-        $startDate = new \DateTime();
-        $report1 = self::$reportTestHelper->generateReport(self::$em, client: $client, startDate: $startDate);
-        $courtOrder->addReport($report1);
-        self::$fixtures->persist($courtOrder)->flush();
-
-        // add a submitted report to the court order
-        $previousReportStartDate = $startDate->modify('-365 days');
-        $submitDate = $previousReportStartDate->modify('+30 days');
-        $report2 = self::$reportTestHelper->generateReport(self::$em, client: $client, startDate: $previousReportStartDate, dateChecks: false);
-        $courtOrder->addReport($report2);
-        self::$reportTestHelper->submitReport($report2, self::$em, submittedBy: $user, submitDate: $submitDate);
-        self::$fixtures->persist($courtOrder)->flush();
+        ['persons' => ['users' => ['lay1' => $user]], 'orders' => [['hw' => ['order' => $courtOrder, 'reports' => [$submitted, $open]]]]] = self::$fixtureService->instantiateScenario(
+            new Scenario(new CourtOrderDescriptor(DeputySet::oneLay(), CourtOrderReportType::OPG104, 1))
+        );
 
         // login to get the token for API calls
-        $token = self::$client->login('successful-court-order-test@opg.gov.uk', 'DigidepsPass1234', self::$deputySecret);
+        $token = self::loginAsDeputy($user->getEmail());
 
         // make the API call
-        $responseJson = self::$client->assertJsonRequest(
+        $responseJson = self::assertJsonRequest(
             'GET',
             "/v2/courtorder/{$courtOrder->getCourtOrderUid()}",
             ['AuthToken' => $token, 'mustSucceed' => true]
@@ -173,7 +100,7 @@ class CourtOrderControllerTest extends AbstractTestController
             if ($report['submitted']) {
                 // previous submitted report
                 $actualSubmitDate = new \DateTime($report['submit_date']);
-                $this->assertEquals($submitDate->format(\DateTime::ATOM), $actualSubmitDate->format(\DateTime::ATOM));
+                $this->assertEquals($submitted->getSubmitDate()?->format(\DateTime::ATOM), $actualSubmitDate->format(\DateTime::ATOM));
                 $this->assertNull($report['un_submit_date']);
             } else {
                 // current report => no submit date or unsubmit date
@@ -186,19 +113,15 @@ class CourtOrderControllerTest extends AbstractTestController
     public function testInviteDeputyActionNoAuthFail()
     {
         $courtOrderUid = '7747628917';
-
-        $courtOrder = self::$fixtures->createCourtOrder($courtOrderUid, CourtOrderType::PFA, CourtOrderKind::Single, 'ACTIVE');
-        self::$fixtures->persist($courtOrder)->flush();
-
-        self::$client->assertEndpointNeedsAuth('POST', "/v2/courtorder/$courtOrderUid/lay-deputy-invite");
+        self::assertEndpointNeedsAuth('POST', "/v2/courtorder/$courtOrderUid/lay-deputy-invite");
     }
 
     public function testInviteLayDeputyActionInvalidPayloadFail()
     {
-        [$user, $courtOrder] = $this->addUserAndCourtOrderAndDeputy('court-order-invite-deputy-payload-test@opg.gov.uk');
+        ['persons' => ['users' => ['lay1' => $user]], 'orders' => [['hw' => ['order' => $courtOrder]]]] = self::$fixtureService->instantiateScenario(Scenario::newSimpleLayScenario(reportType: CourtOrderReportType::OPG104));
 
         // login to get the token for API calls
-        $token = self::$client->login($user->getEmail(), 'DigidepsPass1234', self::$deputySecret);
+        $token = self::loginAsDeputy($user->getEmail());
 
         // make the API call to associate a new co-deputy with the court order
         $postData = [
@@ -206,7 +129,7 @@ class CourtOrderControllerTest extends AbstractTestController
             'role_name' => User::ROLE_LAY_DEPUTY,
         ];
 
-        $responseJson = self::$client->assertJsonRequest(
+        $responseJson = self::assertJsonRequest(
             'POST',
             "/v2/courtorder/{$courtOrder->getCourtOrderUid()}/lay-deputy-invite",
             ['AuthToken' => $token, 'mustFail' => true, 'data' => $postData]
@@ -218,24 +141,19 @@ class CourtOrderControllerTest extends AbstractTestController
     public function testInviteLayDeputyActionSuccess()
     {
         // $deputy has access to court order; $user is the User entity for $invitingDeputy
-        [$user, $courtOrder] = $this->addUserAndCourtOrderAndDeputy('court-order-invite-deputy-test@opg.gov.uk');
-
-        // client
-        $client = self::$fixtures->createClient(settersMap: ['setCaseNumber' => '1122334455']);
-        $courtOrder->setClient($client);
-        self::$fixtures->persist($courtOrder)->flush();
+        ['client' => $client, 'persons' => ['users' => ['lay1' => $user]], 'orders' => [['hw' => ['order' => $courtOrder]]]] = self::$fixtureService->instantiateScenario(Scenario::newSimpleLayScenario(reportType: CourtOrderReportType::OPG104));
 
         // add pre-reg record for deputy to be invited
-        $invitedDeputy = new PreRegistration([
+        $invitedDeputy = self::$fixtureService->persist(new PreRegistration([
             'Case' => $client->getCaseNumber(),
             'DeputyFirstname' => 'Amz',
             'DeputySurname' => 'Sloroz',
-            'DeputyUid' => '123415235',
-        ]);
-        self::$fixtures->persist($invitedDeputy)->flush();
+            'DeputyUid' => self::$fixtureService->getUid(),
+        ]));
+        self::$fixtureService->flush();
 
         // login to get the token for API calls
-        $token = self::$client->login($user->getEmail(), 'DigidepsPass1234', self::$deputySecret);
+        $token = self::loginAsDeputy($user->getEmail());
 
         // make the API call to associate a new co-deputy with the court order
         $postData = [
@@ -245,7 +163,7 @@ class CourtOrderControllerTest extends AbstractTestController
             'role_name' => User::ROLE_LAY_DEPUTY,
         ];
 
-        $responseJson = self::$client->assertJsonRequest(
+        $responseJson = self::assertJsonRequest(
             'POST',
             "/v2/courtorder/{$courtOrder->getCourtOrderUid()}/lay-deputy-invite",
             ['AuthToken' => $token, 'mustSucceed' => true, 'data' => $postData]
